+++ packages/next-drupal-query/CHANGELOG.md
@@ -8,5 +8,5 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-* **@next-drupal/query:** update type docs ([3851197](https://github.com/chapter-three/next-drupal/commit/3851197251d8240f7102f8781ce66ee815e3aa5f))
+* **next-drupal-query:** update type docs ([3851197](https://github.com/chapter-three/next-drupal/commit/3851197251d8240f7102f8781ce66ee815e3aa5f))
 * add next-drupal-query ([627888e](https://github.com/chapter-three/next-drupal/commit/627888ec06138344893385c7b270f79d29822c1a))